Added docstrings to several functions and removed debug comments

+++ src/custom_type_annotations.py
@@ -2,13 +2,12 @@ from typing import NewType
 from typing import List
 from typing import Union
 from typing import NamedTuple
+from decimal import Decimal
 
 
-
-RepetitionTime = NewType('RepetitionTime', int)
+RepetitionTime = NewType('RepetitionTime', Decimal)
 NumberOfSlices = NewType('NumberOfSlices', int)
 SliceScanOrder = NewType('SliceScanOrder', List[int])
 SliceOrdering = NewType('SliceOrdering', Union['interleaved', 'straight'])
 SliceParams = NewType('SliceParams', NamedTuple)
 SliceTimings= NewType('SliceTimings', List[int])
-
